chore: use the same version constraint in both test infra and release.
The `vsce` tool requires the `devDependencies` to have the same vscode
engine version constraint as the released code but the
`devDependencies` here is only used to bootstrap the tests.. but
it does seem that we have any other way around this limitation.

As npm fetches the latest package that satisfies the condition, then
this doesn't change anything important, only that we cannot enforce
a newer vscode library in the tests.